Test programs working.
add_observer in Popuplist failed since even ENTER Was being notified back,
and it was popping up a list. So circular call.

We can't yet use hotkeys in messageboxes for fields, since they
were created before a form existed.
And a form needs a window. Should be create a form without a window and
use that ??