ComboBoxCellRenderer
For lists, to display combo boxes.
Althoght does not make much sense. THe normal list box already has
selection of rows. so its like a list which is now editable and selectable.

This has only checkboxes. How do i pass in text in addition to booleans
given that the method interface of renderers takes one value.