ruby file explorer: added system commands
can exec a command such as glocate or mdfind and see results in explorer and then use View
or cd into those folders

Since listbox truncates content before passing to listcellrenderer
and sometime we may need entire content for some purpose (File.stat) in
this case, we needed to pass row_index also.
I have added that in the repaint call, after column. So needs to be
fixed in other renderers.