Code to stop table from printing a column if exceeding bounds
EXPERIMENTAL CODE: Affects only tables. Now table will not print a column or header
if columns width causes it to exceed width of table.

This means if you have a single column which is wider than table
then it wont print. In Listboxes we always truncate data and pass to
renderer. We could do that here, but it could get messy.