app with splitpanes
attempting a simpler take on splitpanes
that does the calculations on its own so user does not have
to specifiy too much , also it allows block like operations.
all calculations are done in repaint.

See examples/splitp.rb as to the new way of being able to
created splitpanes and objects in blocks.